replace autoIndex to CurrentFileIndexer
