+++ src/utils.ts
@@ -1,12 +1,13 @@
 import { EditorState } from '@codemirror/state';
 import { SyntaxNodeRef } from '@lezer/common';
-import MathPlugin from 'main';
+import { ENVs_MAP } from 'env';
+import MathPlugin, { VAULT_ROOT } from 'main';
 
 // Generic utility functions handing files.
 
-import { App, TFile, getLinkpath, LinkCache, MarkdownView, renderMath, finishRenderMath, TAbstractFile, TFolder, TextFileView, EditorPosition, Loc, CachedMetadata, SectionCache, parseLinktext, resolveSubpath, Notice } from 'obsidian';
+import { App, TFile, getLinkpath, LinkCache, MarkdownView, renderMath, finishRenderMath, TAbstractFile, TFolder, TextFileView, EditorPosition, Loc, CachedMetadata, SectionCache, parseLinktext, resolveSubpath, Notice, Editor } from 'obsidian';
 import { DataviewApi, getAPI } from 'obsidian-dataview';
-import { PLUGIN_NAME } from 'settings';
+import { DEFAULT_SETTINGS, MathSettings, NumberStyle, PLUGIN_NAME, findNearestAncestorContextSettings } from 'settings';
 
 
 export function validateLinktext(text: string): string {
@@ -256,6 +257,83 @@ export function getBlockIdsWithBacklink(path: string, app: App): string[] {
         }
     }
     return ids;
+}
+
+export const MATH_CALLOUT_PATTERN = /\> *\[\! *math *\|(.*)\](.*)/;
+
+export function matchMathCallout(line: string): RegExpExecArray | null {
+    if (line) {
+        return MATH_CALLOUT_PATTERN.exec(line)
+    }
+    return null;
+}
+
+
+export function readMathCalloutSettingsAndTitle(line: string): { settings: MathSettings, title: string } | undefined {
+    const matchResult = matchMathCallout(line);
+    if (matchResult) {
+        let settings = JSON.parse(matchResult[1]) as MathSettings;
+        let title = matchResult[2].trim();
+        return { settings, title };
+    }
+}
+
+
+export function readMathCalloutSettings(line: string): MathSettings | undefined {    // const matchResult = matchMathCallout(editor, lineNumber);
+    let result = readMathCalloutSettingsAndTitle(line);
+    if (result) {
+        return result.settings;
+    }
+}
+
+
+export function readMathCalloutTitle(line: string): string | undefined {    // const matchResult = matchMathCallout(editor, lineNumber);
+    let result = readMathCalloutSettingsAndTitle(line);
+    if (result) {
+        return result.title;
+    }
+}
+
+export function resolveSettings(settings: MathSettings | undefined, plugin: MathPlugin, currentFile: TAbstractFile) {
+    // Resolves settings. Does not overwride, but returns a new settings object.
+    let contextSettings = findNearestAncestorContextSettings(plugin, currentFile);
+    return Object.assign({}, plugin.settings[VAULT_ROOT], contextSettings, settings);
+}
+
+export function formatTitleWithoutSubtitle(settings: MathSettings): string {
+    let env = ENVs_MAP[settings.type];
+
+    let title = '';
+    if (settings.rename && settings.rename[env.id]) {
+        title = settings.rename[env.id] as string;
+    } else {
+        title = env.printedNames[settings.lang as string];
+    }
+    if (settings.number) {
+        let numberString = '';
+        if (settings.number == 'auto') {
+            if (settings.autoIndex !== undefined) {
+                settings.number_init = settings.number_init ?? 1;
+                let num = +settings.autoIndex + +settings.number_init;
+                let style = settings.number_style ?? DEFAULT_SETTINGS.number_style as NumberStyle;
+                numberString = CONVERTER[style](num);
+            }
+        } else {
+            numberString = settings.number;
+        }
+        if (numberString) {
+            title += ` ${settings.number_prefix}${numberString}${settings.number_suffix}`;
+        }
+    }
+    return title;
+}
+
+export function formatTitle(settings: MathSettings): string {
+    let title = formatTitleWithoutSubtitle(settings);
+    if (settings.title) {
+        title += ` (${settings.title})`;
+    }
+    return title;
 }
 
 const ROMAN = ["", "C", "CC", "CCC", "CD", "D", "DC", "DCC", "DCCC", "CM",
